Basic okhttp setup.  (using CLEARTEXT) started gson utility class.

+++ app/src/main/java/hereticpurge/chuckjoker/database/JokeDao.java
@@ -21,7 +21,7 @@ public interface JokeDao {
     long insertJoke(JokeItem jokeItem);
 
     @Delete
-    long deleteSingleJoke(JokeItem jokeItem);
+    int deleteSingleJoke(JokeItem jokeItem);
 
     @Query("DELETE FROM jokeitem")
     void deleteAllJokeItems();
